Input Handling Rework - Game Objects - Drawable/Transformable Distinction - Vector2 - Logger
Completely reworked input handling so we should now support multiple input devices (atm it's just keyboard and gamepad).

Add a Logger class that we can push messages to and see them in an imgui window. This was helpful for seeing what inputs were being detected and what commands were being run due to that input.

Added a vector2 class for ease as having pairs of ints individually was annoying