Demo
Made a very basic demo that has WSAD movement and space to create projectiles. These projectiles have a direction, velocity and lifetime.

Added normalisation to Vector2 as it is useful

Changed the Input to not be event driven as that caused issues when holding multiple keys down at the same time

Added colour modulation to textures so multiple objects can use the same texture and just have them as different colours

Changed Transformable to take floats instead of ints. Even though ints initially made sense as you can't have 0.1 of a pixel, it ruined the movement interpolation as rounding errors caused jerky or no movements.