Listener/Reporter - Tilemap
Changed things to be event driven via notifies instead of by hacks via bool returns etc

Added a tilemap in using Tiled. You can navigate the map.

I hate the command pattern used for the input. I'm going to change it. It just makes no fkin sense as it means either everything has to live in the GameObject interface or it doesn't work.