Input Handling - DeviceInputInfo
Changed input handling to allow for state checking from the device to alter the command.

For instance, the joystick being moved just past the deadzone vs all the way in one direction will alter the move command deltas accordingly.

We might not want this in all cases but it's a nice to have!