Input - Camera - Tilemap
Changed the input bindings for the controller and keyboard so they move the player not the camera.

Made it so cameras can now have attachments with Update calls. The idea is to have attachments such as LERP or Lock On to interact with the camera in a generic way

Cleaned up the Tilemap stuff by adding in new classes and containing code. Still trying to load object layers but that's what this refactor was for. Next step...