Depth Rendering
Added the option to give rendered objects a z index so that we can then do ordered rendering so we can have depth.

Added z_index as a property to tilemap layers but if that doesn't exist, use their layer index as their z index.

The player can now move in the world and be rendered on top of the grass but behind trees.

This is rather simplistic and will likely cause errors later but it's something!